requirements.txt
mysql-connector-python==9.2.0
pandas==2.2.2
gspread==6.1.4
oauth2client==4.1.3
load_dotenv==0.1.0